% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_na.R, R/any_na.R
\name{all_na}
\alias{all_na}
\alias{all_na.default}
\alias{all.na}
\alias{na.all}
\alias{any_na}
\alias{any_na.default}
\alias{na.any}
\title{Tests for missing values}
\usage{
all_na(x)

\method{all_na}{default}(x)

\method{all}{na}(x)

na.all(x)

any_na(x)

\method{any_na}{default}(x)

na.any(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
logical scalar indicating if values are missing.
}
\description{
Test if all values are missing
}
\details{
These are S3 Generics that provide default methods.

\code{all_na} reports if \strong{all} values are missing.

\code{any_na} reports if \strong{any} values are missing.
}
\examples{

  all_na( c( NA, NA, 1 ) )    # FALSE
  all_na( c( NA, NA, NA ) )   # TRUE
  
  df <- data.frame( char = rep(NA_character_, 3), nums=1:3)
  all_na(df)  # FALSE
  
  df <- data.frame( char = rep(NA_character_, 3), nums=rep(NA_real_,3))
  all_na(df)  # TRUE
  
  any_na( 1:10 )           # FALSE
  any_na( c( 1, NA, 3 ) )  # TRUE

}
\seealso{
\itemize{
\item \code{\link[=which.na]{which.na()}}
}

\itemize{
\item \code{\link[=anyNA]{anyNA()}}
}
}
