% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_na.R
\name{n_na}
\alias{n_na}
\alias{na.howmany}
\alias{na.n}
\alias{pct_na}
\alias{na.pct}
\title{Counts how many values are NA}
\usage{
n_na(x)

na.howmany(x)

na.n(x)

pct_na(x)

na.pct(x)
}
\arguments{
\item{x}{object to count how many values are \code{NA}}
}
\value{
\code{n_na} returns an integer. \code{pct_na} returns a numeric value 0-1.
}
\description{
Returns the number of values that are NA
}
\details{
\code{n_na} counts the number of missing values. \code{na.n} is an alias in the dplyr
style.

\code{pct_na} gives the percentage of values that are \code{NA}
}
\examples{
  x <- c( 1, NA, NA, 4:5 )
  n_na(x)
  pct_na(x)

}
