% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.rm.R
\name{na.rm}
\alias{na.rm}
\title{na.rm}
\usage{
na.rm(object, ...)
}
\arguments{
\item{object}{to remove \code{NA}s from}

\item{...}{further arguments special methods could require.}
}
\value{
An object of the same class with all \code{NA} values removed. For
data.frame and data.table objects entire columns are removed if they
contain solely \code{NA} values.
}
\description{
Removes \code{NA} values from objects
}
\details{
For \strong{vectors} this is the same as \code{\link[stats:na.omit]{stats::na.omit()}} or
\code{\link[stats:na.exclude]{stats::na.exclude()}}. It will also work on recursive objects.

This is predominantly maintained for syntactic convenience since a number of
functions have na.omir
}
\seealso{
\itemize{
\item \code{\link[stats:na.omit]{stats::na.omit()}}, \code{\link[stats:na.exclude]{stats::na.exclude()}}
\item \code{\link[=all_na]{all_na()}}
}
}
