% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_explicit_.R
\docType{data}
\name{NA_explicit_}
\alias{NA_explicit_}
\title{NA_explicit_}
\format{An object of class \code{character} of length 1.}
\usage{
NA_explicit_
}
\description{
Default replacement for missing values in categorical vectors.
}
\details{
\code{NA_explicit_} is used as a default replacement for categorical vectors.

It is an active binding to \code{getOptions('NA_explicit_')} and is exported
to the callers namespace.

To change the value of \code{NA_explicit} use:\preformatted{options( NA_explicit = new_value )
}

\code{NA_explicit_} cannot be directly set.
}
\seealso{
\code{\link[=na.replace]{na.replace()}}
}
\keyword{datasets}
