% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_logical.R
\docType{data}
\name{NA_logical}
\alias{NA_logical}
\title{NA_logical}
\format{An object of class \code{logical} of length 1.}
\usage{
NA_logical
}
\description{
NA_logical
}
\details{
This simply creates a NA_logical variable. This is the same as NA
}
\keyword{datasets}
