% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_na.R
\name{which_na}
\alias{which_na}
\alias{na.which}
\alias{which.na}
\title{Identify missing values}
\usage{
which_na(x)

na.which(x)

which.na(x)
}
\arguments{
\item{x}{vector to search for missing values}
}
\value{
\code{integer} of indexes of \code{x} that corerspond to elements
of x that are missing (\code{NA}).  Names of the result
are set to the names of \code{x}.
}
\description{
Return the indexes of those values that are missing
}
\details{
\code{which_na} is implemented as \code{which( is.na(x) )}.
It is a S3 generic function.
}
\examples{

  x <- c( 1, NA, NA, 4:6 )
  which_na(x)
  
  names(x) <- letters[1:6]
  which_na(x)
  
}
\seealso{
\itemize{
\item \code{\link[base:is.na]{base::is.na()}} - for the variant returning logical
}
}
